requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
django_environ==0.4.5
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
environ==1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
pytz==2021.1
psycopg2==2.9.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0